requirements.txt
Flask==2.3.3
numpy==1.26.0
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
transformers==4.34.0
scikit-learn==1.3.1
torch==2.0.1
Werkzeug==2.3.7
